main.py
import os
import tempfile
import asyncio
from typing import List
from pathlib import Path

from astrbot.api.event import filter, AstrMessageEvent, MessageEventResult
from astrbot.api.star import Context, Star, register
from astrbot.api import logger
from astrbot.api.message_components import *

# 引入所需的第三方库
try:
    from pixivpy3 import AppPixivAPI
    import requests
    from PIL import Image
    from reportlab.pdfgen import canvas
    from reportlab.lib.pagesizes import A4
    from reportlab.lib.utils import ImageReader
    import io
except ImportError as e:
    logger.error(f"缺少必要的依赖包: {e}")
    logger.error("请安装: pip install pixivpy3 pillow reportlab requests")

@register("pid2pdf", "Joker42S", "根据Pixiv ID下载图片并保存为PDF发送", "1.0.0")
class Pid2PdfPlugin(Star):
    def __init__(self, context: Context):
        super().__init__(context)
        self.api = None
        self.temp_dir = None

    async def initialize(self):
        """插件初始化方法"""
        try:
            # 初始化Pixiv API
            self.api = AppPixivAPI()
            # TODO: 这里需要配置Pixiv账号信息
            # self.api.auth(username="your_username", password="your_password")
            
            # 创建临时目录用于存储下载的图片
            self.temp_dir = Path(tempfile.mkdtemp(prefix="pid2pdf_"))
            logger.info(f"Pid2Pdf插件初始化完成，临时目录: {self.temp_dir}")
            
        except Exception as e:
            logger.error(f"Pid2Pdf插件初始化失败: {e}")

    @filter.command("pid2pdf")
    async def pid_to_pdf(self, event: AstrMessageEvent):
        """根据Pixiv ID下载图片并生成PDF"""
        try:
            # 解析用户输入的PID
            message_parts = event.message_str.strip().split()
            if len(message_parts) < 2:
                yield event.plain_result("请提供Pixiv ID，格式: /pid2pdf <PID>")
                return
            
            pid = message_parts[1].strip()
            if not pid.isdigit():
                yield event.plain_result("Pixiv ID必须是数字")
                return
            
            yield event.plain_result(f"开始处理Pixiv ID: {pid}，请稍候...")
            
            # 获取作品详情
            artwork_info = await self._get_artwork_info(pid)
            if not artwork_info:
                yield event.plain_result(f"无法获取PID {pid} 的作品信息")
                return
            
            # 下载图片
            image_paths = await self._download_images(artwork_info)
            if not image_paths:
                yield event.plain_result(f"下载PID {pid} 的图片失败")
                return
            
            # 生成PDF
            pdf_path = await self._create_pdf(image_paths, pid, artwork_info)
            if not pdf_path:
                yield event.plain_result(f"生成PDF失败")
                return
            
            # 发送PDF文件
            await self._send_pdf(event, pdf_path, pid)
            
        except Exception as e:
            logger.error(f"处理PID转PDF时出错: {e}")
            yield event.plain_result(f"处理过程中出现错误: {str(e)}")
        finally:
            # 清理临时文件
            await self._cleanup_temp_files()

    async def _get_artwork_info(self, pid: str) -> dict:
        """获取Pixiv作品信息"""
        try:
            # TODO: 实现获取Pixiv作品信息的逻辑
            # result = self.api.illust_detail(pid)
            # if result.illust:
            #     return result.illust
            # return None
            
            # 临时返回模拟数据
            logger.info(f"获取PID {pid} 的作品信息")
            return {"id": pid, "title": "示例作品", "user": {"name": "示例作者"}}
            
        except Exception as e:
            logger.error(f"获取作品信息失败: {e}")
            return None

    async def _download_images(self, artwork_info: dict) -> List[Path]:
        """下载Pixiv图片"""
        try:
            image_paths = []
            
            # TODO: 实现图片下载逻辑
            # if artwork_info.meta_single_page:
            #     # 单图作品
            #     url = artwork_info.meta_single_page.original_image_url
            #     path = await self._download_single_image(url, 0)
            #     if path:
            #         image_paths.append(path)
            # else:
            #     # 多图作品
            #     for i, page in enumerate(artwork_info.meta_pages):
            #         url = page.image_urls.original
            #         path = await self._download_single_image(url, i)
            #         if path:
            #             image_paths.append(path)
            
            logger.info(f"下载了 {len(image_paths)} 张图片")
            return image_paths
            
        except Exception as e:
            logger.error(f"下载图片失败: {e}")
            return []

    async def _download_single_image(self, url: str, index: int) -> Path:
        """下载单张图片"""
        try:
            # TODO: 实现单张图片下载
            # headers = {'Referer': 'https://www.pixiv.net/'}
            # response = requests.get(url, headers=headers, timeout=30)
            # if response.status_code == 200:
            #     file_extension = url.split('.')[-1]
            #     file_path = self.temp_dir / f"image_{index}.{file_extension}"
            #     with open(file_path, 'wb') as f:
            #         f.write(response.content)
            #     return file_path
            # return None
            
            logger.info(f"下载图片 {index}: {url}")
            return None
            
        except Exception as e:
            logger.error(f"下载单张图片失败: {e}")
            return None

    async def _create_pdf(self, image_paths: List[Path], pid: str, artwork_info: dict) -> Path:
        """将图片转换为PDF"""
        try:
            if not image_paths:
                return None
            
            pdf_path = self.temp_dir / f"pixiv_{pid}.pdf"
            
            # TODO: 实现PDF生成逻辑
            # c = canvas.Canvas(str(pdf_path), pagesize=A4)
            # page_width, page_height = A4
            # 
            # for image_path in image_paths:
            #     # 打开图片
            #     img = Image.open(image_path)
            #     img_width, img_height = img.size
            #     
            #     # 计算缩放比例
            #     scale_w = page_width / img_width
            #     scale_h = page_height / img_height
            #     scale = min(scale_w, scale_h) * 0.9  # 留边距
            #     
            #     new_width = img_width * scale
            #     new_height = img_height * scale
            #     
            #     # 居中显示
            #     x = (page_width - new_width) / 2
            #     y = (page_height - new_height) / 2
            #     
            #     # 添加图片到PDF
            #     c.drawImage(str(image_path), x, y, new_width, new_height)
            #     c.showPage()
            # 
            # c.save()
            
            logger.info(f"生成PDF: {pdf_path}")
            return pdf_path
            
        except Exception as e:
            logger.error(f"生成PDF失败: {e}")
            return None

    async def _send_pdf(self, event: AstrMessageEvent, pdf_path: Path, pid: str):
        """发送PDF文件给用户"""
        try:
            # TODO: 实现文件发送逻辑
            # 根据AstrBot API发送文件
            # with open(pdf_path, 'rb') as f:
            #     file_data = f.read()
            #     yield event.file_result(file_data, f"pixiv_{pid}.pdf")
            
            yield event.plain_result(f"PDF生成完成: pixiv_{pid}.pdf")
            
        except Exception as e:
            logger.error(f"发送PDF失败: {e}")
            yield event.plain_result(f"发送PDF文件失败: {str(e)}")

    async def _cleanup_temp_files(self):
        """清理临时文件"""
        try:
            if self.temp_dir and self.temp_dir.exists():
                import shutil
                shutil.rmtree(self.temp_dir)
                logger.info("清理临时文件完成")
        except Exception as e:
            logger.error(f"清理临时文件失败: {e}")

    @filter.command("pid_help")
    async def help_command(self, event: AstrMessageEvent):
        """显示插件帮助信息"""
        help_text = """
Pid2Pdf 插件使用说明：

命令格式：
/pid2pdf <Pixiv_ID> - 根据Pixiv ID下载图片并生成PDF

示例：
/pid2pdf 123456789

功能：
1. 获取指定Pixiv作品的所有图片
2. 下载图片到本地
3. 将图片合并为PDF文件
4. 发送PDF文件给用户

注意：
- 需要有效的Pixiv账号配置
- 处理时间取决于图片数量和大小
- 大型作品集可能需要较长时间处理
        """
        yield event.plain_result(help_text.strip())

    async def terminate(self):
        """插件销毁方法"""
        await self._cleanup_temp_files()
        logger.info("Pid2Pdf插件已销毁")
