requirements.txt
# Pixiv API客户端
pixivpy3>=3.7.0

# 图片处理
Pillow>=9.0.0

# PDF生成
reportlab>=3.6.0

# HTTP请求
requests>=2.28.0

# 异步支持 (如果需要)
aiohttp>=3.8.0

# 路径处理（Python 3.4+内置，但明确列出用于兼容性）
pathlib2>=2.3.0; python_version<"3.4" 